requirements.txt
fastapi
pydantic
uvicorn
python-multipart
openai-whisper
pyannote.audio
torch
numpy
